// src/app/automation/configuration/action/createaction/types/uicontrol/models/UIConditionType.ts
export enum UIConditionType {
  visible = 'VISIBLE',
  exist = 'EXIST',
  hidden = 'HIDDEN',
  appear = 'APPEAR',
  disappear = 'DISAPPEAR',
  readonly = 'READONLY',
  empty = 'EMPTY',
  image = 'IMAGE',
  focused = 'FOCUSED',
  enabled = 'ENABLED',
  disabled = 'DISABLED',
  selected = 'SELECTED',
  checked = 'CHECKED',
}